+++ lib/control/src/lib.rs
@@ -1,6 +1,6 @@
 #![cfg_attr(not(test), no_std)]
 
 pub mod config;
+pub mod controllers;
 pub mod hp_relay;
-pub mod lev_controllers;
 pub mod pneumatics;
